Don't insit on inlining Function* llvmcalls
Since they can drag with them a significant number of other functions,
it is easier to just call them. In the future, we'll probably want to
run an LLVM inling pass, which will restore any benefit this had for
performance. Still keep the inlining for the text version of llvmcall
though in case people rely on it.