Make sure `other` is not referenced anywhere in jl_get_llvmf
When regenerating the LLVM IR, make the jlcall wrapper use the declaration
pointing to the original function rather than the newly generated function.
That way it is safe to delete `other` in jl_get_llvmf. In the future it
would be good to have a function that just recompiles the wrapper without
always compiling both, but that is out of the scope of this PR.